BREAKING Scraper: Put burden of sorting search results on the scraper plugin
Because the identified metadata is not guaranteed to return the game title, nor is the scraper plugin obligated to use such data, it is up to the scraper plugin to determine which game is the best match for the game. Plugins may use whatever metadata is available to them to order the list.