GameCacheServer: Implement image scaling and access to pre-scaled images on GameCacheServer
The syntax is as follows:

For pre-scaled images (@10, @25, @50, @75), simply prepend the respective scale string to the end of the image name, ex. BoxartFront@25 will return the prescaled 25% sized image. If multiple scales strings are appended, only the smallest will be processed.

For screenshots that do not pre-scale, and to scale GameMediaCache images on-the-fly, append the scale factor as the GET parameter 'scale' to the URL. For example, ?scale=50 will scale the image half. To prevent memory leaks GameCacheServer will only scale down images, and never scale over 100%.