BREAKING: Make GameDatabase backed by Dapper micro-ORM.
This makes IGameDatabase -> IGameLibrary, which represents an abstracted view of the previous GameDatabase. The concrete GameDatabase class remains to interface with Dapper and DapperExtensions that map the games in the database to actual classes, while the concrete GamesLibrary class wraps the GamesDatabase, allowing it's internals to be reimplemented or rewritten to adapt to other databases. GameDatabase, and other classes relating to databases are now hidden away as internal; the introduction of an ORM resulted in a lot more boilerplate code, but as a trade off should be more tested and robust than manually writing in SQL statements. This also introduces a GameInfo internal constructor for serialization use only and a reorganizing of their arguments.