FileSystem: Introduce the ManifestedDirectory API
The ManifestedDirectory will serve as the basis of the new FileRecords.
Instead of FileRecords being kept in a database, the file will be
assigned a GUID kept in a folder-basis manifest file, that will be
reference-able on the database to keep metadata.

GUIDs are kept with the file, so files that are moved around will
keep metadata, so long as movement is kept within the API.